% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominant_eigenvector.R
\name{domiant_eigenvector}
\alias{domiant_eigenvector}
\title{Find the domiant eigenvector of a class}
\usage{
domiant_eigenvector(X, w)
}
\arguments{
\item{X}{A dataset or data matrix consisting of single class entries}

\item{w}{The adaboost weights vector, corresponding to how difficult a particular
instance is to classify correctly.}
}
\value{
a unit vector u corresponding to the direction to construct the householder
matrix about.
}
\description{
Finds the (weighted) dominant eigenvector from one of the two
classes in the dataset, this is used to constuct the dataset reflection.
}
