% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenboost.R
\name{predict.eigenboost}
\alias{predict.eigenboost}
\title{Create predictions from Eigenboost model}
\usage{
\method{predict}{eigenboost}(object, new_data, n_tree = NULL)
}
\arguments{
\item{object}{An object of type eigenboost}

\item{new_data}{A data frame or data matrix with which to create new predictions}

\item{n_tree}{Only take the first n_tree many trees in the ensemble to create new prediction}
}
\value{
Returns the predicted class in c(-1, 1)
}
\description{
Aggregates predictions from all trees in the ensemble.
}
