% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenboost_print.R
\name{print.eigenboost}
\alias{print.eigenboost}
\title{Print an Eigenboost object}
\usage{
\method{print}{eigenboost}(x, ...)
}
\arguments{
\item{x}{An eigenboost model object}

\item{...}{Not used}
}
\description{
Pretty print the values of the eigenboost model object.
}
